+++ src/databricks/labs/ucx/source_code/notebooks/magic.py
@@ -13,15 +13,6 @@ from databricks.labs.ucx.source_code.linters.python_ast import NodeBase, Tree
 class MagicCommand(NodeBase):
 
     @classmethod
-    def convert_magic_lines_to_magic_commands(cls, python_code: str):
-        lines = python_code.split("\n")
-        for i, line in enumerate(lines):
-            if not line.startswith("%"):
-                continue
-            lines[i] = f"magic_command({line.encode()!r})"
-        return "\n".join(lines)
-
-    @classmethod
     def extract_from_tree(
         cls, tree: Tree, problem_factory: ProblemFactory
     ) -> tuple[list[MagicCommand], list[DependencyProblem]]:
